+++ src/perks/year_5_perks.rs
@@ -12,7 +12,7 @@ use super::{
 
 pub fn fmr_cascade_point(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -38,13 +38,13 @@ pub fn fmr_cascade_point(
 
 pub fn sbr_encore(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
 ) -> HashMap<u32, i32> {
     let mut map = HashMap::new();
-    let val = clamp(_value, 0, 4);
+    let val = clamp(_value, 0, 4) as i32;
     let stability_boost = 8 * val;
     let range_boost = 5 * val;
     map.insert(StatHashes::RANGE.to_u32(), range_boost);
@@ -54,7 +54,7 @@ pub fn sbr_encore(
 
 pub fn dmr_focused_fury(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -77,7 +77,7 @@ pub fn dmr_focused_fury(
 
 pub fn rmr_fragile_focus(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -93,7 +93,7 @@ pub fn rmr_fragile_focus(
 
 pub fn sbr_fragile_focus(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -109,7 +109,7 @@ pub fn sbr_fragile_focus(
 
 pub fn dmr_gutshot_straight(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -134,7 +134,7 @@ pub fn dmr_gutshot_straight(
 
 pub fn sbr_offhand_strike(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -150,7 +150,7 @@ pub fn sbr_offhand_strike(
 
 pub fn rmr_offhand_strike(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -169,7 +169,7 @@ pub fn rmr_offhand_strike(
 
 pub fn hmr_slickdraw(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -183,7 +183,7 @@ pub fn hmr_slickdraw(
 
 pub(super) fn sbr_stats_for_all(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -208,7 +208,7 @@ pub(super) fn sbr_stats_for_all(
 
 pub(super) fn hmr_stats_for_all(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -227,7 +227,7 @@ pub(super) fn hmr_stats_for_all(
 
 pub(super) fn rmr_stats_for_all(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -248,7 +248,7 @@ pub(super) fn rmr_stats_for_all(
 
 pub(super) fn rsmr_stats_for_all(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -268,7 +268,7 @@ pub(super) fn rsmr_stats_for_all(
 
 pub(super) fn sbr_steady_hands(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -284,7 +284,7 @@ pub(super) fn sbr_steady_hands(
 
 pub(super) fn hmr_steady_hands(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -305,7 +305,7 @@ pub(super) fn hmr_steady_hands(
 
 pub(super) fn dmr_target_lock(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -345,7 +345,7 @@ pub(super) fn dmr_target_lock(
 
 pub(super) fn dmr_over_under(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
@@ -365,12 +365,12 @@ pub(super) fn dmr_over_under(
 
 pub(super) fn sbr_well_rounded(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
 ) -> HashMap<u32, i32> {
-    let val = clamp(_value, 0, 2);
+    let val = clamp(_value, 0, 2) as i32;
     let mut map = HashMap::new();
     let stat_base = if _is_enhanced { 12 } else { 10 };
     let stat_bump = stat_base * val;
@@ -382,12 +382,12 @@ pub(super) fn sbr_well_rounded(
 
 pub(super) fn hmr_well_rounded(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
 ) -> HandlingModifierResponse {
-    let val = clamp(_value, 0, 2);
+    let val = clamp(_value, 0, 2) as i32;
     //due to ease of activation and upkeep will assume its always active
     // let mut duration = if _is_enhanced {9.0} else {8.5};
     let stat_base = if _is_enhanced { 12 } else { 10 };
@@ -401,12 +401,12 @@ pub(super) fn hmr_well_rounded(
 
 pub(super) fn rmr_well_rounded(
     _input: &CalculationInput,
-    _value: i32,
+    _value: u32,
     _is_enhanced: bool,
     _pvp: bool,
     _cached_data: &HashMap<String, f64>,
 ) -> RangeModifierResponse {
-    let val = clamp(_value, 0, 2);
+    let val = clamp(_value, 0, 2) as i32;
     let stat_base = if _is_enhanced { 12 } else { 10 };
     let range = stat_base * val;
     RangeModifierResponse {
